requirements.txt
auth~=0.5.3
Flask~=3.0.3
flask-mysql